WIP: Split usage & definition
